refactoring cli structure
