+++ c8ylp/timer.py
@@ -1,0 +1,52 @@
+"""Timer"""
+
+import logging
+import time
+from datetime import timedelta
+from typing import Callable
+
+
+class CommandTimer:
+    """Command Timer which shows how long a command takes to run
+    and prints out a message to the user
+
+    Example
+
+    >>>
+    with CommandTimer():
+        print("Doing someting")
+        time.sleep(100)
+    >>>
+
+    """
+
+    def __init__(self, message: str, on_exit: Callable[[], None] = None) -> None:
+        self.message = message
+        self.start_time = 0
+        self.last_duration = 0
+        self._on_exit = on_exit
+
+    def start(self):
+        """Start the timer"""
+        self.start_time = time.monotonic()
+
+    def stop(self) -> float:
+        """Stop the timer and return the duration in seconds
+
+        Returns:
+            float: Duration in seconds
+        """
+        if not self.start_time:
+            return 0
+        self.last_duration = time.monotonic() - self.start_time
+        return self.last_duration
+
+    def __enter__(self) -> None:
+        self.start()
+
+    def __exit__(self, _type, _value, _traceback) -> None:
+        duration = timedelta(seconds=(int(self.stop())))
+        msg = f"{self.message}: {duration}"
+        logging.info(msg)
+        if callable(self._on_exit):
+            self._on_exit(msg)
